Extract shared dataset generation helpers into utility modules
